Added file detection and user-assignable CSV elements
mt.load(True) tells the code it is in debug mode and automatically assigns values, however, mt.load() or mt.load(False) will allow the importing of different CSV files. The user will also be asked what the label title is as well as the descriptor data so they can have differently named CSV files. I also moved stripping to a seperate file to do something similar, however, it also allows uses the given titles from the intial setup to strip the data (minus the genre bit, that requires a discussion with Chris to decide on when this value is given)